[libmspack-xenia] lzxd.c: Change `position_slots` from `unsigned int` to `unsigned short`.
Very minor space optimization. `position_slots` is only referenced once
during lzxd initialization, so this shouldn't cause a performance issue.

Code size differences: (64-bit Gentoo Linux, gcc-12.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
  14091       0       0   14091    370b ./extlib/libmspack-xenia/CMakeFiles/mspack.dir/lzxd.c.o
  14053       0       0   14053    36e5 ./extlib/libmspack-xenia/CMakeFiles/mspack.dir/lzxd.c.o
    -38       0       0     -38     -26 Difference

I would have expected the difference to be 22, since position_slots has
11 elements, but I guess reducing the size may have reduced some padding
as well.

+++ extlib/libmspack-xenia/lzxd.c
@@ -211,7 +211,7 @@ static int lzxd_read_lens(struct lzxd_stream *lzx, unsigned char *lens,
  * position_base[0] = 0
  * position_base[i] = position_base[i-1] + (1 << extra_bits[i-1])
  */
-static const unsigned int position_slots[11] = {
+static const unsigned short position_slots[11] = {
     30, 32, 34, 36, 38, 42, 50, 66, 98, 162, 290
 };
 static const unsigned char extra_bits[36] = {
